Improved 'configure.sh' with installation/update logic
This commit enhances the `configure.sh` script. It now handles both initial installation and updates of the `asus_fanctl` program. 
If any local changes are made to the .c script, simply run the configure.sh script to push updated binary to /usr/local/bin

It detects if an existing binary is present,  stops the service, removes the old binary, recompiles from the updated source code, and restarts the service.